fix: made export builtin norm compliant
