+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: svereten <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 16:24:55 by svereten          #+#    #+#             */
-/*   Updated: 2025/01/31 14:25:55 by svereten         ###   ########.fr       */
+/*   Updated: 2025/01/31 14:31:20 by svereten         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #ifndef MINISHELL_H
@@ -116,6 +116,7 @@ void		signal_int(int signal);
 int			builtin_cd(t_cmd *cmd);
 int			builtin_env(t_cmd *cmd);
 int			builtin_export(t_cmd *cmd);
+int			builtin_export_no_args(void);
 void		builtin_exit(t_cmd *cmd);
 	
 void		minishell_exit(int status, char *msg);
