+++ src/builtins/export/no_args.c
@@ -1,17 +1,15 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   export.c                                           :+:      :+:    :+:   */
+/*   no_args.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: svereten <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/01/31 13:45:52 by svereten          #+#    #+#             */
-/*   Updated: 2025/01/31 14:27:07 by svereten         ###   ########.fr       */
+/*   Created: 2025/01/31 14:31:44 by svereten          #+#    #+#             */
+/*   Updated: 2025/01/31 14:34:21 by svereten         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "minishell.h"
-#include "command.h"
-#include <stdio.h>
 
 /**
  * "declare -x " strlen is 11 + '\n' makes 12
@@ -39,6 +37,19 @@ static size_t	export_get_len(void)
 	return (len);
 }
 
+static void	export_concat_output(t_env_var *cur, char *output, size_t len)
+{
+	ft_strlcat(output, "declare -x ", len + 1);
+	ft_strlcat(output, cur->key, len + 1);
+	if (cur->value)
+	{
+		ft_strlcat(output, "=\"", len + 1);
+		ft_strlcat(output, cur->value, len + 1);
+		ft_strlcat(output, "\"", len + 1);
+	}
+	ft_strlcat(output, "\n", len + 1);
+}
+
 static char	*export_get_output(void)
 {
 	t_env_var	*cur;
@@ -57,55 +68,22 @@ static char	*export_get_output(void)
 			cur = cur->next;
 			continue ;
 		}
-		ft_strlcat(output, "declare -x ", len + 1);
-		ft_strlcat(output, cur->key, len + 1);
-		if (cur->value)
-		{
-			ft_strlcat(output, "=\"", len + 1);
-			ft_strlcat(output, cur->value, len + 1);
-			ft_strlcat(output, "\"", len + 1);
-		}
-		ft_strlcat(output, "\n", len + 1);
+		export_concat_output(cur, output, len);
 		cur = cur->next;
 	}
 	return (output);
 }
 
-static int	export_no_args(void)
+/**
+ * Why not `printf`?
+ * Same logic as env builtin
+ */
+int	builtin_export_no_args(void)
 {
 	char	*output;
 
 	output = export_get_output();
 	if (ft_putstr_fd(output, STDOUT_FILENO) < 0)
 		minishell_exit(1, NULL);
-	return (0);
-}
-
-int	builtin_export(t_cmd *cmd)
-{
-	int			i;
-	char		*eq_sign_loc;
-	char		*key;
-	t_env_var	*node;
-
-	if (!cmd->argv[1])
-		return (export_no_args());
-	i = 1;
-	while (cmd->argv[i])
-	{
-		eq_sign_loc = ft_strchr(cmd->argv[i], '=');
-		if (!eq_sign_loc)
-			ft_get_alloc_env_node(cmd->argv[i]);
-		else
-		{
-			key = ft_substri(cmd->argv[i], 0, eq_sign_loc - cmd->argv[i]);
-			if (!key)
-				minishell_exit(1, NULL);
-			node = ft_get_alloc_env_node(key);
-			node->value = ft_substri(cmd->argv[i], eq_sign_loc - cmd->argv[i], ft_strlen(cmd->argv[i]) - 1);
-			ft_free(STR, &key);
-		}
-		i++;
-	}
 	return (0);
 }
